Packages: upgrade webpack, babel and jest-cli
This upgrades us to the latest versions of webpack and babel.
We had to upgrade to babel v24 because earlier versions were using
babel-core v6.

This commit reworks the webpack and babel configs and switches out the
`extract-text-webpack-plugin` because it was replaced with
`mini-css-extract-plugin`.

I was able to confirm that the built CSS was unchanged. The JS was
slightly changed because webpack has changed its runtime.

TODO: We might want to switch webpack to production mode.